refactor(datafusion): rename catalog provider modules
