+++ crates/integrations/datafusion/src/lib.rs
@@ -15,14 +15,14 @@
 // specific language governing permissions and limitations
 // under the License.
 
-mod catalog;
-pub use catalog::*;
+mod catalog_provider;
+pub use catalog_provider::*;
 
 mod error;
 pub use error::*;
 
 pub mod physical_plan;
-mod schema;
+mod schema_provider;
 pub mod table;
 pub use table::table_provider_factory::IcebergTableProviderFactory;
 pub use table::*;
