+++ crates/integrations/datafusion/src/schema_provider.rs
